Super tests - mama, look, I am coding like a pro

+++ src/main/java/models/MakeTransferResponse.java
@@ -5,13 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 public class MakeTransferResponse extends BaseModel{
     private int receiverAccountId;
-    private float amount;
+    private BigDecimal amount;
     private String message;
     private int senderAccountId;
